requirements.txt
# Core Data Science
numpy>=1.24.0
pandas>=2.0.0
xarray>=2023.0.0
netCDF4>=1.6.0
dask>=2023.0.0

# Hybrid ML (The "Brain")
scikit-learn>=1.3.0  # KMeans, Random Forest
groq>=0.4.0          # Ultra-Fast AI (Llama 3)

# Visualization ("The Dashboard")
streamlit>=1.25.0
pydeck>=0.8.0
plotly>=5.15.0
matplotlib>=3.7.0

# Reports & Audio
geopy>=2.4.0
reportlab>=4.0.0
gTTS>=2.4.0

# Connectivity
requests>=2.31.0
google-generativeai # Keeping as backup
